Separate tests from main code; add coverage; misc project resources
