+++ .gitignore
@@ -1,0 +1,8 @@
+*.egg-info
+__pycache__
+
+.pytest_cache
+.coverage
+htmlcov/
+
+.DS_Store
