+++ README.md
@@ -1,0 +1,13 @@
+# FROST
+
+BIP340 compatible implementation of Flexible Round-Optimized Schnorr Threshold Signatures (FROST).
+
+This work is made possible with the support of Brink.
+
+* [Brink blog: Jesse Posner on FROST](https://brink.dev/blog/2021/04/15/frost/)
+* [Introductory slides](FROST.pdf)
+* [FROST flow diagram](dot/api/frost.pdf)
+
+
+### Running the tests
+see: [tests/README.md](tests/README.md)
