+++ pyproject.toml
@@ -1,0 +1,20 @@
+[project]
+name = "FROST"
+version = "0.0.1"
+authors = [
+    {name = "Jesse Posner"},
+]
+description = "BIP340 compatible implementation of Flexible Round-Optimized Schnorr Threshold Signatures (FROST)."
+readme = "README.md"
+requires-python = ">=3.10.2"
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "License :: OSI Approved :: MIT License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3.10",
+]
+
+[tool.coverage.run]
+source = ["src"]
