+++ tests/requirements.txt
@@ -1,0 +1,2 @@
+coverage==7.3.4
+pytest==7.4.3
